+++ GPTifier/src/interface/command_chats.cpp
@@ -6,12 +6,51 @@
 #include "utils.hpp"
 
 #include <fmt/core.h>
+#include <getopt.h>
 #include <string>
 #include <vector>
 
 namespace {
 
 // List chats -----------------------------------------------------------------------------------------------
+
+ParamsGetChatCompletions read_cli_list_cc(int argc, char **argv)
+{
+    ParamsGetChatCompletions params;
+
+    while (true) {
+        static struct option long_options[] = {
+            { "help", no_argument, 0, 'h' },
+            { "json", no_argument, 0, 'j' },
+            { "limit", required_argument, 0, 'l' },
+            { 0, 0, 0, 0 },
+        };
+
+        int option_index = 0;
+        int c = getopt_long(argc, argv, "hjl:", long_options, &option_index);
+
+        if (c == -1) {
+            break;
+        }
+
+        switch (c) {
+            case 'h':
+                cli::help_command_chats_list();
+                exit(EXIT_SUCCESS);
+            case 'j':
+                params.print_raw_json = true;
+                break;
+            case 'l':
+                params.limit = optarg;
+                break;
+            default:
+                cli::exit_on_failure();
+        }
+    };
+
+    params.sanitize();
+    return params;
+}
 
 void print_chat_completions(const ChatCompletions &ccs)
 {
@@ -29,7 +68,7 @@ void print_chat_completions(const ChatCompletions &ccs)
 
 void command_chats_list(int argc, char **argv)
 {
-    ParamsGetChatCompletions params = cli::get_opts_get_chat_completions(argc, argv);
+    ParamsGetChatCompletions params = read_cli_list_cc(argc, argv);
     ChatCompletions ccs = get_chat_completions(std::get<int>(params.limit));
 
     if (params.print_raw_json) {
